pyproject.toml
[project]
name = "ai-quiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.3",
    "httpx>=0.28.1",
    "openai>=1.57.2",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.2",
]
